+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -551,6 +551,7 @@ class Configuration implements ConfigurationInterface
                         ->scalarNode('version_strategy')->defaultNull()->end()
                         ->scalarNode('version')->defaultNull()->end()
                         ->scalarNode('version_format')->defaultValue('%%s?%%s')->end()
+                        ->scalarNode('manifest_path')->defaultNull()->end()
                         ->scalarNode('base_path')->defaultValue('')->end()
                         ->arrayNode('base_urls')
                             ->requiresAtLeastOneElement()
@@ -567,6 +568,12 @@ class Configuration implements ConfigurationInterface
                         })
                         ->thenInvalid('You cannot use both "version_strategy" and "version" at the same time under "assets".')
                     ->end()
+                    ->validate()
+                        ->ifTrue(function ($v) {
+                            return isset($v['version_strategy']) && $v['version_strategy'] == 'json_manifest' && !isset($v['manifest_path']);
+                        })
+                        ->thenInvalid('You must configure the "manifest_path" key in order to use the "json_manifest" "version_strategy".')
+                    ->end()
                     ->fixXmlConfig('package')
                     ->children()
                         ->arrayNode('packages')
@@ -582,6 +589,7 @@ class Configuration implements ConfigurationInterface
                                         ->end()
                                     ->end()
                                     ->scalarNode('version_format')->defaultNull()->end()
+                                    ->scalarNode('manifest_path')->defaultNull()->end()
                                     ->scalarNode('base_path')->defaultValue('')->end()
                                     ->arrayNode('base_urls')
                                         ->requiresAtLeastOneElement()
@@ -597,6 +605,12 @@ class Configuration implements ConfigurationInterface
                                         return isset($v['version_strategy']) && isset($v['version']);
                                     })
                                     ->thenInvalid('You cannot use both "version_strategy" and "version" at the same time under "assets" packages.')
+                                ->end()
+                                ->validate()
+                                    ->ifTrue(function ($v) {
+                                        return isset($v['version_strategy']) && $v['version_strategy'] == 'json_manifest' && !isset($v['manifest_path']);
+                                    })
+                                    ->thenInvalid('You must configure the "manifest_path" key in order to use the "json_manifest" "version_strategy".')
                                 ->end()
                             ->end()
                         ->end()