+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/assets.php
@@ -24,6 +24,10 @@ $container->loadFromExtension('framework', array(
                 'base_urls' => array('https://bar2.example.com'),
                 'version_strategy' => 'assets.custom_version_strategy',
             ),
+            'json_manifest_strategy' => array(
+                'version_strategy' => 'json_manifest',
+                'manifest_path' => '/path/to/manifest.json'
+            ),
         ),
     ),
 ));