+++ src/Symfony/Component/Asset/VersionStrategy/JsonManifestVersionStrategy.php
@@ -1,0 +1,73 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Asset\VersionStrategy;
+
+use Symfony\Component\Asset\Exception\MissingAssetManifestException;
+
+/**
+ * Reads the versioned path of an asset from a JSON manifest file.
+ *
+ * For example, the manifest file might look like this:
+ *     {
+ *         "main.js": "main.abc123.js",
+ *         "css/styles.css": "css/styles.555abc.css"
+ *     }
+ *
+ * You could then as for the version of "main.js" or "css/styles.css".
+ */
+class JsonManifestVersionStrategy implements VersionStrategyInterface
+{
+    private $manifestPath;
+
+    private $manifestData;
+
+    public function __construct($manifestPath)
+    {
+        $this->manifestPath = $manifestPath;
+    }
+
+    /**
+     * With a manifest, we don't really know or care about what
+     * the version is. Instead, this returns the path to the
+     * versioned file.
+     *
+     * @param string $path
+     * @return string
+     */
+    public function getVersion($path)
+    {
+        return $this->applyVersion($path);
+    }
+
+    public function applyVersion($path)
+    {
+        $manifestPath = $this->getManifestPath($path);
+
+        return $manifestPath ? $manifestPath : $path;
+    }
+
+    private function getManifestPath($path)
+    {
+        if (null === $this->manifestData) {
+            if (!file_exists($this->manifestPath)) {
+                throw new \RuntimeException(sprintf('Asset manifest file "%s" does not exist.', $this->manifestPath));
+            }
+
+            $this->manifestData = json_decode(file_get_contents($this->manifestPath), true);
+            if (0 < $errorCode = json_last_error()) {
+                throw new \RuntimeException(sprintf('Error parsing JSON from asset manifest file "%s" - %s', $this->manifestPath, json_last_error_msg()));
+            }
+        }
+
+        return isset($this->manifestData[$path]) ? $this->manifestData[$path] : null;
+    }
+}