+++ src/Symfony/Component/Asset/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+3.3.0
+-----
+ * JsonManifestVersionStrategy was added as a way to read final,
+   versioned paths from a JSON manifest file.
+
 2.7.0
 -----